Updated Rakefile
Updated the install task to check if the target symlink/file already
exists. If it does, then do nothing

+++ Rakefile
@@ -32,7 +32,9 @@ end
 desc "create all the symlinks"
 task :install do
   linkables do |linkable, file, target|
-    `ln -s "$PWD/#{linkable}" "#{target}"`
+    unless File.exists?(target)
+      `ln -s "$PWD/#{linkable}" "#{target}"`
+    end
   end
 end
